```             inl rec loop data i =                 match i with                 | () -> ()                 | i :: i' ->                     match i with                     | {from=from'} ->                         inl check near_to' = view from' near_to'                         match i with                         | {near_to=near_to'} -> check near_to'                         | {by} -> check (from' + by)                         | _ -> view from' near_to                     | () -> view 0 near_to                     | from' -> rest from' ```
Yeah, this is where I need to start. Let me just write a comment on that guy I linked to.

Basically the critical piece his entire idea rests on - namely a way of accurately assessing individual characteristics in order to assign values is highly difficult. If it was possible then optimization would not be such a difficult field. Solving credit assignment to any degree of satisfaction is not something humans will be able to do. His idea is only workable if you see aristocrats as being 130 IQ and the masses 100 on average - this is roughly the frame he implies. The future will be more like 1,000,000 and 100. I am still yet to see anyone seriously argue that NPC should literally be NPCs, but things will slowly get to that point.

And my own desire will be instrumental in getting there. Nonetheless, I find his writing a breath of fresh air.

Humanity is definitely reaching the critical limits of bullshit and with this implication I fully agree with him.

Let me make the `view`, `rest` and `check` functions. I need to refocus.