I got up 30m ago and have been resting since then. Note to self - next time leave a note in the latest commit that I am moving branches. I had no idea that I had people looking over these commits. It is not good to cause needless concern.
Concerns, huh? Yeah, I do have some. Despite the recent progress in ML I do not really get the feeling that I am grasping the essence of differentiable programming. It feels more like I am learning tricks. Because of that, the whole plan of making a good poker player is at risk.

I've been trying to brainstorm ideas for the past six months, but they are not generally coming up well. I get the sense that while in the right direction, my shots are missing the mark.

The other point of concern is hardware - GPUs just suck for RL period.

It is hard for me to admit this, but because of this my plan is really not going in a direction of a decisive victory.

Well, whatever the case, I'll know by the end of next year. Hopefully by that time I will have done everything I wanted to and will know firmly where I stand. I understand well enough that a single breakthrough could change everything and despite all my disadvantages, poker is simple enough that I just might slip through the crack to the other side where the victory lies.

For now, let me rest a bit more and then I will start programming.