Let me stop being so lazy and do some work.
I am sure that true Inspired could just flip a switch and they would find the power to work 24/7, but I am just a human so the most I can do is put a little bit of effort every day. Today won't be any different.

Yesterday was such a bog standard programming day to which I need to return to. I thought a bit, programmed and then called it a day. That kind of pace is ideal. Let me do it today as well.