"10:10am.
```
tns 3 4 5 .get |> Console.writeline
```

The previous test worked, but when I do this I get a can't apply the tensor anymore error. Is this just a problem in views or is it present in the original tensor?

10:15am.

```
inl tns =
    HostTensor.init (2,3,4) (inl a b c -> a*b*c)
    //|> ViewHostTensor.wrap ({from=2; near_to=4},{from=2; near_to=5},{from=2; near_to=6})

tns 1 2 3 .get |> Console.writeline
```

Only in the view it seems. I'll have to track it down from there then."