``` let test113 =     "test113",[view_host_tensor;console],"Do the tensor views work?",     """ inl tns =     HostTensor.init (2,3,4) (inl a b c -> a*b*c)     |> ViewHostTensor.wrap ({from=2; near_to=4},{from=2; near_to=5},{from=2; near_to=6})
inl tns = tns ((), {from=3; by=2}, {from=3})
tns .basic |> HostTensor.print
    """
```

This test is perfection. As expected, apart from a few type errors that were easy to fix, everything works perfectly. The next thing to test would be tree views. They are the cause for this redesign - without them I would need to calculate views by hand. They are well worth putting into the library for that reason.

One of the reason why Cuda programming is so dangerous in C is because of the need to calculate indices for multidimensional arrays by hand. It is such an embarrassment on Nvidia's side and I am glad that I have the privilege of spearheading the adoption of a better way of doing it.

Though GPU programming is really something that will have to die before it gets better. The imperative way of doing it is too firmly lodged in by now with their group for Spiral to make any difference. Had I started work on this a decade earlier I might have made a difference, but it is just too late now. C++ won't be going anywhere.

...Let me test the tree view. I am getting too sidetracked.