Today the chores took a while, but let me do some work in whatever time is left.
First let me replace the HostTensor module in the standard library with the new one and add the tests along with the HostTensorView module.

Then I will rediff the cache. Somehow I get the feeling I will run into that problem where it spontaneously shows changes for everything. That will be troublesome, but I will deal with somehow.

I suspect that the hacker_rank tests which use views won't compile properly. I'll have to modify them to use the views.

Well, let me make that my mission for the rest of the day. Tomorrow I will tackle the much larger task of dealing with the Cuda using libraries.