After doing a quick review, I've come to the conclusion that I do not need to do any work in particular at all. Maybe I'll remove all those uses of length though.
There is one place on block reduce that I know uses ranges explicitly, but that is just one thing I need to fix. I guess this redesign won't be as hard as I thought it would be...but actually, at least for `CudaKernel` functions I should aim to preserve the range functionality.

...This will actually be less work than doing the otherwise. Well, great.

Let me call it a day here.