https://arxiv.org/abs/1810.00303 Newton-MR: Newton's Method Without Smoothness or Convexity
https://www.reddit.com/r/MachineLearning/comments/9l7j46/d_why_do_machine_learning_papers_have_such/?st=jmw2iucu&sh=7adeddb7

Let me rest a little with this paper and link. I said I would call it a day, but it is too early for that. That having said, I do not feel like programming anymore. My spells are spent and I need to recharge.

I also need to dig up that list of bullet points. Let me do so now.

```
~- Remove tanh in the `generalize_mi` for the the `plastic_rnn`. I did not mean for this to be in. I am going to have to repeat the experiments for this.~
- Make `fused_activation`.
~- Extend host tensor to support named views.~
- Extend `oja_update` to support multiple updates at once.
~- Extend the module syntax to support expressions inside of them. Parenthesis will be used on the left side of with/without for that.~
- Add the `matrix` class in the initializer.
- Add `stream` in the initializer.
```

Ah, here it is. This redesign that I just did was a big one, no doubt about that. I'll do the other 3 tomorrow and maybe finish the plastic RNN.

Also since I am doing full optimization, I am going to go with the old stream idea to speed up the matrix multiplies. Last time I got irritated that I could not do this dynamically, but this time I'll be more mature about it. I've added it to the bullet points list.

I think that just like with the redesign, perfecting these optimizations will take a few days at least, but in the end it will be worth it. I need to go through this exercise once. It is not that expect this plastic RNN to last me long - the important thing I am aiming for is to make all the future optimizations of this type easier.

It is a challenge I need to overcome using Spiral.

Once I do this, I'll have conquered the optimized LSTM challenge from 2016 by proxy and will have cleaned up all the loose ends from 2016. Well, apart from having convolutional operations, but let me do one thing at a time. I do not need those yet.