There is a certain sense of finality in what I've been doing here.
Strictly speaking, back in 2015 I could have just gone the Python + Tensorflow route and I would have been long done with the poker part by now.

But I was obstinate. I wanted to prove my deep belief in the new way and this was my way of doing it. But that will be over soon. I'll soon be at the stage where I could have been long, long ago - not just 2015, but I could have been at this stage over a decade ago. The long chain of mistakes will be finally broken.

Yeah, yeah, I know. I could have gone with Theano, Tensorflow or PyTorch, but that is the false path.

Let me take PyTorch to show what I mean.

'It is currently in the process of adding JIT compilation.'

This single sentence sums up my opposition to both it and to the opposing frameworks.

PyTorch is more than a restricted DSL, it is an autograd library. Meaning the JIT would have to optimize AD while supporting full Python features.

This is bound to fail and those frameworks will fall far short of their performance goals in the end. So far, Tensorflow cannot do this even with static graphs. Anticipating their failure, along with the need to make programming on future hardware easier is why I've started the Spiral project.

Attempts to make dynamic languages on par with C, and to optimize monadic computation in functional languages have all failed. The kinds of optimizations fast AD requires are just too hard for optimizing compilers. They need to be done by hand in a fashion.

Literally the only choice is to make a language with more sensible semantics and draw out the necessary performance that way. This was my choice with Spiral. It is the only sane sensible choice, but if I left it to others I would just get something like Julia or Rust - two languages that really miss the point of partial evaluation and static typing.

Ah, what a burden. I am convinced that at this point the world really is ruled by memes. There is really only so much one man can do on his own.

But this world went a bit too far with this nonsense - if truth becomes evil then evil will simply have to win. I won't accept any other outcome. Perception is not reality - otherwise I am an utter fool in putting so much effort in shifting my mind to get a better grasp on it.

Lately, I've been getting sparks of the same feeling I had in 2013 and 2014, back then when I first figured out the self improvement method I lovingly named Necromutation. It is absolutely nothing concrete yet, but probably my subconcious is starting to grasp something. It will probably bubble its way to the surface eventually. Though it seems impossible, there should exist a much more powerful self improvement method than the one I described a few years ago in story Simulacrum - though from the context I'd guess that the price for it will even be more severe than for Necromutation.