This is the continuation of the previous commit. Last night I had some time to think about how I am going to do the optimized activations. I'll ignore the bullet points about adding stuff to the initializer - those are trivial and straightforward.
Lately, I've been thinking about CurveBall and second order optimization. To be honest, I did not quite mean to make the library such a huge bet on first order optimization. While doing the review, after I decided how I am going to do the activation I could see for the first time just how many unique parts the library has. It is quite amazing. I was really a busy bee for the past 10 months.

The way I am going to do the activations starting from here is to use the LSTM one as a model. I could have done it like this from the start, but it is amazing that it would not have helped much - for example, a lot of the work that went into the cost functions that deals with them in their optimized form. Second order AD would just wreck this.

Nevertheless, for standard stuff like the Oja update, and the fused activations, I am definitely going to be using kernels that support AD as they would be to hard to do otherwise. The time to add this to the arsenal has been long overdue. Moreover, being able to do this in Spiral is something that the framework JITs won't be able to come close to. It is pretty ironic that I am adding AD for the sake of efficiency.

I'll will make doing the optimized RNNs in Spiral a step harder than normal ones. That will be an accomplishment. In F# it would be past the threshold of complexity, and in C++ I won't even consider it. PyTorch is an absolute embarrassment for having that extending with C++ segment.