Let finish the series with this third commit.
Though I have the idea of where I want to go now, I feel like lazing for a bit in order to brainstorm the details. I need to gather the urge to do it first. I am going to have to sketch out the optimized RNN again, now that I have the new tensor and its views. Furthermore, I cannot do the optimizations all in tandem. Today I should narrow my focus and do just one of them.

The one that I should do it should be the matrix multiply fuses. Then the fused activations would come tomorrow and the streams the day after that. It may take shorter or longer than that depending on how things go. I am not in a rush. This is the prelude to the beginning which will start off with me doing the very, very, very basic thing I should have done long ago - just using the AC on poker.