Let me at least try to do a little today. I'll start work on a new branch 'initializer_rewrite'. Honestly, I know the feeling of burn out well - if I really wanted to program that badly I would do it instead of lying in bed for over four hours and bolstering my will by thinking of fox girls. This won't be the first nor the last time I will go through this. I definitely badly want to just get through this so I can move to doing RL, but it is important to realize when there is a large gaps that needs to be crossed first so I do not keep falling into potholes. Just because I want to stop doing design work does not mean I should. I need to do it perfectly. That is the mantra that got me all the way to this point.
I had enough time to think about it. The way the current initializer is structured is just as bad as the previous one, except in different ways.

What I will do is forget about the fused multiplies for a bit. Now that the new tensors is here this is the ideal time to redesign the initializer from the ground up. I know how to do it and where I went wrong last time. What I am also going to this time is add checkpointing. The time has finally come to be able to save and load models from files.

It is kind of ironic, but what I am going to do in the initializer is make heavy use of OO programming. I am also going to disallow the use of block when doing declarations. That means no more `wrap` for cells, but I do not think I am using that in the current library anyway.

Killing several birds with one stone - that is the benefit of good design.

Let me make a new list. I am going to go through these in the order they are given rather then last time which was ordered randomly.

- Redesign the initializer
- Do the optimized activations
- Make use of streams in the matrix multiply
- Figure out how to make proper use of K-FAC without bolting it to the matrix multiply. Switch to using it during the optimization stage.
- Make use of reversible learning in the plastic RNN

Each of these points has its own challenges that I am going to have to go through. But once I go through all of them I will have the infrastructure in place to deal with the optimization of any RNN that might arise in the future. I am not going put stress on myself to get this done quickly.

Given how long the tensor redesign took me, depending on how things go this might take me until the end of the month. Just forget about RL for the time being. It is time for me to roll up my sleeves and do some dirty work once more. But this is exactly the kind of work I've been doing for the past 3.5 years. The 3 weeks of past month when I only played with DP was a mirage. If I want to earn the privilege of getting back to it and getting to poker I am going to have to earn it by perfecting the RNN.

That is the way things must go.

In my experience, the mountain becomes easier to climb once you stop pretending that it is a pebble. I need to do this for its own sake and for Spiral. I never *needed* to any of this for RL to begin with. This is a bet separate myself from the crowd and ensure my future. This is so I can improve my own programming.