I decided to remove all the Oja and the PRONG stuff. I removed most of the library thanks to that. I'll do it all from the ground up. I guess I do not necessarily have to start with the plastic RNN. I could always do the optimizations on the MI-RNN to start with.
https://arxiv.org/abs/1810.02334
Unsupervised Learning via Meta-Learning

Oh, I did not think a paper like this would come out so soon. I'll read it later.

Let me commit here. So it is decided - I'll forget about the plastic RNN for a bit and focus on optimizing the MI-RNN. I think I should add the LSTM to the mix. It is a great baseline and a good target for optimization that I shold just have it. In contrast the plastic RNN is more challenging as its perfected form requires reversing the weight updates. I'd like to put that off until the end as it will be challenging to pull off.

...Yeah, forget the MIRNN. Let me just go with the LSTM for now. I'll change my goal for the first 4 points to optimizing the LSTM. The MIRNN will be trivial once I do the LSTM. Plastic RNN will come after that.

I need to work out the details of how exactly do I want the initializer to look like now.