Yeah, it is the same thing.
```
    inl generalized_mi = // Online only version for now.
        activation {
            fwd=inl {input state bias={si s i c}} -> si * state * input + s * state + i * input + c
            bck=inl {in={input state bias={si s i c}} out} ->
                {
                input = si * state + i
                state = si * input + s
                bias = { si = input*state; i = input; s = state; c = one }
                }
            }

    inl generalized_mi_tanh = // Online only version for now.
        activation {
            fwd=inl {input state bias={si s i c}} -> si * state * input + s * state + i * input + c |> tanh_fwd
            bck=inl {in={input state bias={si s i c}} out} ->
                inl out = tanh_bck out
                {
                input = si * state + i
                state = si * input + s
                bias = { si = input*state; i = input; s = state; c = one }
                } |> Struct.map ((*) out)
            }
```

~- Redesign the initializer~
- Do the optimized activations
- Make use of streams in the matrix multiply
- Figure out how to make proper use of K-FAC without bolting it to the matrix multiply. Switch to using it during the optimization stage.
- Make use of reversible learning in the plastic RNN

Now come the activations. Strictly speaking, for the MIRNN, these activations are already optimized. What i mean to say is to make them easier to create.

For complex functions having to do the backward pass by hand is absolutely an error. I need to get these linear optimizations out of the way for good. I felt that when I was experimenting with the Hebbian learning I was really pushing my luck there.

Here I'll take advantage of atomics to simplify things and implement them as they should be.

Let me take a short break first.