```     // Optimized LSTM activation.     inl lstm memory_old {ins with input_cell output_cell forget_cell memory_cell} s =         inl b,a as dim = primal memory_old .dim         inl primals_ins = primals ins         assert_dim primals_ins (Struct.map (const dim) primals_ins)         inl in = {ins with memory_old}
        inl out =
            inl x = to_dev_tensor {in=primals in}
            s.CudaFun.init {dim} (inl dim -> CudaAD.activation_lstm dim x .out |> primals)
            |> HostTensor.unzip
            |> Struct.map (dr s)

        {
        out
        bck=met _ ->
            inl x = to_dev_tensor {in out}
            s.CudaKernel.iter {dim} (inl dim -> CudaAD.activation_lstm dim x |> CudaAD.run |> ignore)
        }
```

It is good that this is still here. I can use this as the basis for the new `map`.

I really do not feel like doing it right now though. The reason for that is that I'd need to think about the design for the AD module.

Well, at least for starters, let me pull the `lstm_activation` out of it. What I am going to have to do is get rid of those `link_adjoints` and have them be done automatically. For a one off function what I did was fine, but a general map would need higher standards. Let me back it up here.

```
inl activation_lstm dim x =
    inm {from with in={input_cell output_cell memory_cell forget_cell memory_old}} = link dim x

    inm memory =
        inm input = sigmoid input_cell
        inm forget = sigmoid forget_cell
        inm memory = tanh memory_cell

        inm a = input * memory
        inm b = forget * memory_old
        a + b
    inm _ = try_link_adjoint dim {from to={out={memory}}}

    inm out =
        inm memory = tanh memory
        inm output = sigmoid output_cell
        output * memory
    inm _ = try_link_adjoint dim {from to={out={out}}}

    succ {memory out}
```

How about I start by working on this. Then I will sketch out the `map`.