``` inl activation_lstm dim x =     inm memory =         inm input = sigmoid input_cell         inm forget = sigmoid forget_cell         inm memory = tanh memory_cell
        inm a = input * memory
        inm b = forget * memory_old
        a + b

    inm out =
        inm memory = tanh memory
        inm output = sigmoid output_cell
        output * memory

    succ {memory out}
```

This is what fundamentally this function should be. Nice and easy implementation. No plumbing that should be in the primitive instead.

Right now, I am just getting distracted. I really am tired. Let me see if I can put a dent into the new map function and then I will call it a day. I'll sketch it out in the CudaAD module. I'll leave filling in the details for tomorrow.