"5pm.
```
    inl init {dim} init s =
        inl out =
            s.CudaFun.init {dim} (inl dim -> primals (init dim))
            |> HostTensor.unzip
            |> Struct.map (dr s)

        {
        out
        bck=met _ ->
            inl out = to_dev_tensor out
            s.CudaKernel.iter {dim} (inl dim -> init dim |> link_adjoint out |> CudaAD.run |> ignore)
        }
```

It is amazing how much the clarity and the generality of the code improves when you aim for generality. Yeah, I did not need that `try_link` thing at all. I should have just done it like this from the start.

5:05pm.

```
inl link_adjoint dim {from to} =
    Struct.iter2 (inl from to ->
        match from, to with
        | {adjoint=from}, {adjoint=to} -> to 0 <- from 0
        | _ -> ()
        ) from to
    {
    out=()
    bck=inl _ ->
        Struct.iter2 (inl from to ->
            match from, to with
            | {adjoint=from}, {adjoint=to} -> from 0 <- to 0
            | _ -> ()
            ) from to
    }
```

Ah, so this is how it works.

Yeah, this is not going to work. I need to reverse the order of monadic execution. This `bck` needs to come last. That will need to be done by hand.

```
inl try_link_adjoint dim {from to} =
    inl a = module_intersect (inl from to -> from) from to
    inl b = module_intersect (inl from to -> to) from to

    Struct.map2 (inl from to -> link_adjoint dim {from to} |> inl x -> {x with block=()}) a b
    |> sequence
```

Let me get rid of this. Why the hell did I even make `module_intersect`? I remember it taking some effort.

Also, now that I am going to be using these kinds of activations, it is time to consider putting the `__restrict__` next to type declarations in order to help the optimizer reorder the loads better. This would be an hour of work at most in the codegen. I'll leave it for later.

...Hmmm, what is that `dim` argument even doing? It seems I am not even using it.

Well, let me implement this and then I will call it a day."