``` inl link {dim cur} x =     inl out =         Struct.map <| function             | {primal adjoint} -> primal cur .get |> dr             | x -> x cur .get
    {
    out
    bck = inl _ ->
        Struct.iter2 (inl x out ->
            match x, out with
            | {adjoint=x}, {adjoint=out} ->
                inl x = x cur
                inl out = out 0
                if x.dim = dim then x .set (x .get + out)
                else atomic_add x out
            | _ -> ()
            ) x out
    }
```

This is how the link should be. I do not need to get the adjoints for the inputs explicitly. Why keep them around?

Thanks to this I will be able to easily support tensors smaller than `dim` in `init`.

```
inl map in f s =
    inl in = zip in |> to_dev_tensor
    inl dim = in.dim
    init {dim} (inl cur -> link {dim cur} in >>= f) s
```

Here is map once more. Though I whined that I do not want to program anymore earlier, I really got into it just now.

This is actually very good work that I did today. It is definitely a step beyond what I did when I made the lstm activation. With this I could just take that lstm activation function and pass it into map without a problem.

Having a `init` and `map` of capable doing inline AD in a library is a definite asset for Spiral. If the need arises I might need to do the same for `init_seq`, but apart from layer norm I do not need it right now.

Damn, this is good. Good programming really does warm my heart. At this rate, I might really get somewhere.

My goal for tomorrow will be putting this into action and testing it out. I think that I did everything I wanted for today.