```     inl init s size dsc =         Struct.foldl_map (inl sublayer_size {x with init} ->             inl {d with dsc size} = init sublayer_size             inl weights = Struct.map' (inl x -> x s |> inl x -> x.init; x) dsc |> heap             {x without init with weights size}, size             ) size dsc
    inl run s input =
        Struct.foldl_map (inl input {layer with apply size} ->
            inl input =
                inl input = {input size}
```

I think should suffice to get `size` in there.

I do not have to resolve the name collision. The `init` in Primitives comes earlier. Ok.

---

Ugh. Though I do not feel like it, I have to it. There is nobody else but me who can do it. It is time for the testing gauntlet once again.

Once I do that I'll be done with the second bullet point. Apart from the last one which is kind of special, the other bullet points will fall naturally once I am done with this.

I swear on my desire, I will take a break in the form of doing RL once I am done with all of this. I will do the Maze and the Cue tests once I am done with all of this and just play around. I will do poker. I will bring in self play and async training. I will conquer poker.

I just did so much in 2018. So, so much. This will be the year RL will finally fall into place.

Once that is done I will have a proper power. I think what I need to do past that is do the basics that I missed.

I've been thinking about it for a while, but even though I've been doing ML all this time, I've been really neglecting basic studies of it. I've been thinking even though I've ruled it out as impossible previously, it might be possible to reason through what ML is doing to a limited degree. I need to tap into it.

I am simply going to need improved reasoning capabilities if I am going to move forward from here. The field does its stupid proofs which are like source code in some alien language, but somehow I feel the urge to distill and simplify them for myself.

Those numerical stability tests for the reversible learning updates are the first real bit of science that I had ever done. I need to do more along those lines. Once I cast away my fear and sharpen my mind, the truth will come to me.

But that comes later.

Really, making money in poker is quite simple - all I need is a good agent and then I need to find a way to plug it into those sites and the waters will flow. This was always summary of it.

Now that I have the plastic RNN and PRONG, should that suffice the first part of the plan will fall naturally into place. The second part I will hack together somehow. It will require some engineering effort, but I will find a way to deal with that. After all, I had plenty of practice in that sort of thing.

There is no reason why I could not finish the first part of the plan by the end of the next month.

Yeah, let me do this and I will get to it. It cannot possibly be the case that I will just be working on the ML library forever. At some point it is going to cross the the threshold of usefulness.

Let me do it. The future me is going to have so much fun.

For my own future, I have to do clear the tedious work on my plate today, so that I might get to the fun days that lie ahead.

I'll get into it once I start.