I did not do much in the last hour. I got impatient so I decided to have lunch rather than force myself to program. I might as well call it a day here.
Ahhh...the days of my life go by so quickly. What I am doing here might seem trivial, but it is a ladder to reach a higher place. I might not be doing ML research, but this library will give me capabilities the other ones do not have. Making these flexible map functions will make it much easier to add higher order optimization should the need arise. But the most important thing would be that it will allow me to experiment safely with novel architectures without sacrificing either speed or safety in any way.

The use of atomics will sacrifice reproducibility a bit though, but that is fine. RL is stochastic anyway so I won't even notice.

Right now I need to meditate a bit on how exactly do I want to design broadcasting. I am not entirely sure how to fit the pieces together. Such work is best done when the mind is fresh. I'll give is some thought overnight.

Tomorrow I will finish this and then work on adding streams into both the initializer and the matrix multiply.

Right now I am not entirely done with the first bullet point just yet. I am missing the `tensor_view`, `stream` and `constant`.

I'll add things to it as I need. By the time I am done with the LSTM and PRONG everything will be in it.

When I wrote that it might take me until the end of the month I was heavily lethargic and wanted to release the burden on myself, but at the current pace I will be done with the first 4 bullet points in 3 days or so. Good.

These redesigns always take a lot of time, there is just no way around it. They lead to a lot of code being abandoned, but it is usually code I do not need. Fundamental things tend to take root in both the language and the library while the transients go away.

For this effort, will I be rewarded or will I be made a fool?

Back in 2016, it just seemed so ridiculous for me to claim that I am seeking the Singularity when I cannot make even a generic Cuda map function? Bringing about the apocalypse seems just so ridiculous when the language I was using barely let me express myself. That shame drove me to create what I have now. In terms of power Spiral is an entirely different league than F#. I have my chance.

I will never stop. Even if I burn out, I will rise like a phoenix the next moment ready to try again.