"10:55am.
```
The learning rate is 2 ** -11.00083
Starting timing for: iteration 0
The time was 00:00:06.1215947 for: iteration 0
Training: 2.6166457595468
The learning rate is 2 ** -11.00083
Starting timing for: iteration 1
The time was 00:00:06.0913372 for: iteration 1
Training: 2.04930005404715
The learning rate is 2 ** -11.00083
Starting timing for: iteration 2
The time was 00:00:06.2505851 for: iteration 2
Training: 1.887025096036
The learning rate is 2 ** -11.00083
Starting timing for: iteration 3
The time was 00:00:06.1625488 for: iteration 3
Training: 1.7972763438216
The learning rate is 2 ** -11.00083
Starting timing for: iteration 4
The time was 00:00:06.2228514 for: iteration 4
Training: 1.74115919866873
```

It works. I had to lower the learning rates though.

That having said, 6s is kind of slow for this isn't it?

11am. Hmmm, well, let me do a litmus test."