"``` inl generalized_mi_tanh {bias={si s i c} input state} = one //si * state * input + s * state + i * input + c |> tanh ```
Let me test it with this.

```
inl generalized_mi_tanh {bias={si s i c} input state} = succ one //si * state * input + s * state + i * input + c |> tanh
```

```
Error trace on line: 127, column: 28 in file "CudaAD".
    Struct.iter2 (inl from to -> to 0 <- from dim .get) from to
                           ^
Expected a heap mutable module, reference or an array the input to mutable set.
Got: []
```

Let me fix this.

```
inl link_adjoint dim {from to} =
    Struct.iter2 (inl from to ->
        match to with
        | () -> ()
        | _ -> to 0 <- from dim .get
        ) from to
    {
    out=()
    bck=inl _ ->
        Struct.iter2 (inl from to ->
            match to with
            | () -> ()
            | _ -> from dim .set (to 0)
            ) from to
    }
```

All the checks make it ugly, but whatever.

11:05am.

```
The learning rate is 2 ** -14.97738
Starting timing for: iteration 0
The time was 00:00:06.2008820 for: iteration 0
Training: 3.39248400514343
The learning rate is 2 ** -14.97738
Starting timing for: iteration 1
The time was 00:00:06.1144866 for: iteration 1
Training: 3.3534617210081
The learning rate is 2 ** -14.97738
Starting timing for: iteration 2
The time was 00:00:06.1095803 for: iteration 2
Training: 3.35157219562953
The learning rate is 2 ** -14.97738
Starting timing for: iteration 3
The time was 00:00:06.1888997 for: iteration 3
Training: 3.35078896327382
The learning rate is 2 ** -14.97738
Starting timing for: iteration 4
The time was 00:00:06.0724722 for: iteration 4
Training: 3.35035559469286
```

Hmm, it is actually suspiciously slow now.

11:10am. No, it is fine.

I remember the MIRNN being around 4.2 with the wavefront and 5.2 without it in the previous library, but that does not matter right now. Maybe fiddling with compiler options will help. That can come later.

Good work me.

With this the second bullet point is 85% complete. I still want to make the Oja update using this, but for now it is perfectly fine. It will be trivial to implement with either `map` or `init`.

11:15am. Ah, before I take a break let me put the check in map into `HostTensor`."