``` /// Asserts that all dimensions of the tensors are broadcastable. Returns the dimensions of the largest combination of tensors if applicable. /// tensor structure -> (dim | ()) inl assert_broadcastable =     Struct.foldl (inl s x ->         match s with         | () -> x.dim         | _ ->             assert (eq_type s x.dim) "The inputs must have the same number of dimensions."             Struct.map2 (inl s x ->                 if s = 1 then x                 else                     assert (s = x) "The dimensions of the inputs must all be either singular or equal to each other."                     s                 ) s x.dim         ) () ```
Let me do it like this. ...Yeah, it compiles. Great.

There is plenty I can be pleased about here.

The next thing on the list would be to add streams to the initializer and then take advantage of that in the matrix multiply. Once I have that I'll move to implementing the LSTM. Everything is going as per plan.

Let me take a break here.