It is time to resume this thing.
To be honest, I am a bit unmotivated, so let me just focus on what I would said I would do which is adding streams to the initializer and then I'll put some thought into how I will deal with matmultb.

Actually, I am lying.

I know exactly what I need to do, I just do not want to think about how I am going to redesign matmultb. It is going to be a pain in the ass, much like the broadcasting activation was yesterday.

It is going to be one of those simple things that takes a lot of brainpower and time.

One thing that annoys me right now is having to figure out how exactly I am going to parse the arguments for it.

Let me just give it unambigous semantics and later I will make a proper parser for the damn thing. I do not want to agitate myself over it.

Having to deal with unambiguity is the biggest annoyance for me anywhere.

Let me just do this and I will call it a day. It is time to strike. After I am done with all of this, I'll be able to move on to more fun things.