"5:10pm. It does have them. Good. No parse errors either.
I remember the time where I considered it a badge of honor to work until 8pm, but I've since then learned to not take stress lightly. It is the kind of thing that willpower should not be used to plow through especially once it starts manifesting a physical pain. Is is better to err on the side of caution and stop early - especially if that early is 5-6pm. The number of good working hours in a day are really much fewer than I would like them to be.

I can't wait for the bright post-Singularity future where all work will be play without limits. It really broke me in 2014 once I realized what fool I have been and just how distant were the desires that I had from the desires that I needed to move me in the direction of power. I had no choice but to give myself a kick in order to start moving in the correct direction.

It is pity that my trading days ended on such a sour note, but I can console myself with the fact that the speculation skill that I acquired is what is now guiding me towards the goal that I seek.

Once I make a good player, I think I will make a private project where I will be able to work on the interfaces. I'll leave the language, the ML library and the players public, but I'll have to keep some things a trade secret. Honestly, getting to the point where I work on the interfaces seemed really dramatic when I first envisioned it, but always when the work is done it seems banal afterward. This time it will be the same. Ultimately the plan boils down to 'run RL algorithms on poker.' Natural gradients and plastic RNNs are still a subset of that no matter how I embellish it.

Tomorrow, what I will do is finish the MIRNN and the LSTM. No doubt I will have to do a bunch of design still on the `matmult_stream` - maybe I will add the biases as well. Linear layers could use that. For nonlinear ones, I would be better off fusing the biases with activations.

I'll do the MIRNN and the LSTM in the most straightforward way possible. Matrix multiply and then activation. Two steps are all that I need.

Then after I am done with this tomorrow, comes K-FAC.

I felt pretty smart when I figured out the backwards pass trick, but it seems keeping it till the end is a better idea after all. Even for the poker game, the disadvantage of using short sequences that I am doing now will be offset by the gains from parallelism."