https://arxiv.org/abs/1810.03642 CAML - Fast Context Adaptation via Meta-Learning
Huh, it is this late? Holy crap.

I have no idea how I managed to sleep in this late.

I wanted to say that I was warming up to the idea of putting in higher order optimization for the sake of implementing CurveBall and that I was going to spend some time reading the paper again, but now I am not sure what to do.

Well, there is no point in rushing this. Let me just read the papers that I want and then I'll take a break. I'll program later in the day then.